+++ documentation/changelog.rst
@@ -131,8 +131,10 @@ New features
 Infrastructure / Support
 ----------------------
 
+- Add benchmark script to test speed of the current code base, running some basic commands. [see `PR #787 <https://github.com/FlexMeasures/flexmeasures/pull/787>`_]
 - Introduce a new one-to-many relation between assets, allowing the definition of an asset's parent (which is also an asset). This hierarchical relationship enables assets to be related in a structured manner. [see `PR #855 <https://github.com/FlexMeasures/flexmeasures/pull/855>`_ and `PR #874 <https://github.com/FlexMeasures/flexmeasures/pull/874>`_]
 - Introduce a new format for the output of ``Scheduler`` to prepare for multiple outputs [see `PR #879 <https://github.com/FlexMeasures/flexmeasures/pull/879>`_]
+
 
 
 v0.16.1 | October 2, 2023